simpler config parsing, refact'd report generation
The config parsing actions are now part of the "App::Watson::Configuration"
class. The various actions are now functions of their own.

The report generation now defers the outputting to reduce side effects. It
was also refactored to use separate functions for files and dirs, cutting down
on loops.

The $VERSION back-and-forth was fixed after consulting the "version" module and
<www.cpan.org/modules/04pause.html#version>.

Minor style improvements across the board.